Refactored the Feature class
- One file per class
- Changed exception types from base exception to more specific types
