+++ source/Xunit.Gherkin.Quick/ScenarioDiscoverer.cs
@@ -1,0 +1,56 @@
+﻿using Gherkin;
+using Gherkin.Ast;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Reflection;
+using Xunit.Abstractions;
+using Xunit.Sdk;
+
+namespace Xunit.Gherkin.Quick
+{
+
+    public sealed class ScenarioDiscoverer : IXunitTestCaseDiscoverer
+    {
+        private readonly IMessageSink _messageSink;
+
+        public ScenarioDiscoverer(IMessageSink messageSink)
+        {
+            _messageSink = messageSink;
+        }
+
+        public IEnumerable<IXunitTestCase> Discover(ITestFrameworkDiscoveryOptions discoveryOptions, ITestMethod testMethod, IAttributeInfo factAttribute)
+        {
+            var gherkinDocument = GetGherkinDocumentByType(testMethod.TestClass.Class.ToRuntimeType());
+
+            foreach (var scenario in gherkinDocument.Feature.Children)
+            {
+                yield return new ScenarioXUnitTestCase(_messageSink, testMethod, $"{gherkinDocument.Feature.Name} :: {scenario.Name}", new object[] { scenario.Name });
+            }
+        }
+
+        public static GherkinDocument GetGherkinDocumentByType(Type classType)
+        {
+            var fileName = classType.FullName;
+            fileName = fileName.Substring(fileName.LastIndexOf('.') + 1) + ".feature";
+
+            if (!System.IO.File.Exists(fileName))
+            {
+                var path = (classType.GetTypeInfo().GetCustomAttributes(typeof(FeatureFileAttribute))
+                    .FirstOrDefault() as FeatureFileAttribute)
+                    ?.Path;
+
+                if (path == null || !System.IO.File.Exists(path))
+                {
+                    throw new TypeLoadException($"Cannot find feature file `{fileName}` in the output root directory. If it's somewhere else, use {nameof(FeatureFileAttribute)} to specify file path.");
+                }
+
+                fileName = path;
+            }
+
+            var parser = new Parser();
+            var gherkinDocument = parser.Parse(fileName);
+            return gherkinDocument;
+        }
+    }
+}
